+++ src/commands/cache.rs
@@ -85,7 +85,7 @@ pub async fn cmd_cache_prune(output_json: bool) -> Result<()> {
             .ok()
             .and_then(|m| m.modified().ok())
             .and_then(|t| t.duration_since(std::time::UNIX_EPOCH).ok())
-            .map(|d| d.as_nanos() as i64)
+            .map(|d| d.as_nanos())
         {
             Some(ns) => ns,
             None => continue,
@@ -174,7 +174,7 @@ pub async fn cmd_cache_rebuild(output_json: bool) -> Result<()> {
                     .ok()
                     .and_then(|m| m.modified().ok())
                     .and_then(|t| t.duration_since(std::time::UNIX_EPOCH).ok())
-                    .map(|d| d.as_nanos() as i64)?;
+                    .map(|d| d.as_nanos())?;
 
                 let title = ticket.title.as_deref().unwrap_or("");
                 let body = ticket.body.as_deref().unwrap_or("");
